// src/models/matriculaModel.ts
export interface MatriculaType {
    nocontrol: string;
    curp: string;
    lugarNacimiento: string;
    nombre: string;
    primerApellido: string;
    segundoApellido: string;
    seguro: string;
    genero: string;
    celular: string;
    correo: string;
    indigena: string;
    carrera: string;
    modalidad: string;
    estudios: string;
    entidad: string;
    municipio: string;
    nombreUReal: string;
    nombreUOficial: string;
    discapacidad: string;
    semestre: string;
    status: string;
}